+++ app/src/main/java/trianglz/models/Subcategory.java
@@ -22,7 +22,7 @@ public class Subcategory {
     @SerializedName("quizzes_total")
     public double quizzesTotal;
     @SerializedName("quizzes_grade")
-    public int quizzesGrade;
+    public double quizzesGrade;
     @SerializedName("quizzes")
     public ArrayList<Quizzes> quizzes;
     @SerializedName("assignments_total")
@@ -34,7 +34,7 @@ public class Subcategory {
     @SerializedName("grade_items_total")
     public double gradeItemsTotal;
     @SerializedName("grade_items_grade")
-    public int gradeItemsGrade;
+    public double gradeItemsGrade;
     @SerializedName("grade_items")
     public ArrayList<GradeItems> gradeItems;
     @SerializedName("percentage")
